Common: extract UPnP namespace from NetPlayServer
